% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{hello world expamle 2.0}
\usage{
hello(x = NULL)
}
\arguments{
\item{x}{value to be printed}
}
\description{
liuSHHFSDIUBEKJvnhfbFDVNSDCBHSDBJDSC
JKSBDchjBSDJvNHCVGC
}
\details{
jhbfchjndskjnmkdnvbcbsdhcbjksbljhcb
slkmxc
}
\examples{

hello()

}
