% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ln.R
\name{ln_complex}
\alias{ln_complex}
\title{linear reservir}
\usage{
ln_complex(
  IN,
  a = 0.25,
  S0_1 = 100,
  k_1 = 0.1,
  S0_2 = 15,
  k_2 = 0.2,
  S0_3 = 50,
  k_3 = 0.05
)
}
\arguments{
\item{IN}{input}

\item{a}{alpha}

\item{S0_1}{storage}

\item{k_1}{dischrage}

\item{S0_2}{storage}

\item{k_2}{dischrage}

\item{S0_3}{storage}

\item{k_3}{dischrage}
}
\description{
linear reservir
}
\examples{

ln_complex(IN = rnorm(n = 100))

}
