% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_comma.R
\name{str_split_comma}
\alias{str_split_comma}
\title{Split a string with commas}
\usage{
str_split_comma(string)
}
\arguments{
\item{string}{A character vector with, at most, one element with commas as the delimiter.}
}
\value{
A character vector.
}
\description{
Split a string with commas
}
\examples{
x <- "alfa,bravo,charlie,delta"
str_split_comma(x)
}
