% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_nums.R
\name{str_split_nums}
\alias{str_split_nums}
\title{Split a string and return individual numbers}
\usage{
str_split_nums(string, pattern)
}
\arguments{
\item{string}{A character vector with, at most, one element with only numbers besides the commas.}

\item{pattern}{Pattern to look for.

The default interpretation is a regular expression, as described in
\code{vignette("regular-expressions")}. Use \code{\link[stringr:regex]{regex()}} for finer control of the
matching behaviour.

Match a fixed string (i.e. by comparing only bytes), using
\code{\link[stringr:fixed]{fixed()}}. This is fast, but approximate. Generally,
for matching human text, you'll want \code{\link[stringr:coll]{coll()}} which
respects character matching rules for the specified locale.

Match character, word, line and sentence boundaries with
\code{\link[stringr:boundary]{boundary()}}. The empty string, \verb{""``, is equivalent to }boundary("character")`.}
}
\value{
A numeric vector that is split up.
}
\description{
Split a string and return individual numbers
}
\examples{
x <- "1,2,3"
str_split_num(x)
}
