% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Chop up a string.}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character vector with only element/word}

\item{split}{THe delimiter to split x on.}
}
\value{
A character vector with multiple elements
}
\description{
Chop up a string.
}
\examples{
x <- "zulu; yankee; x-ray; whisker"
strsplit1(x, split = ";")

}
